refactor import torch
