+++ cornac/models/ncf/pytorch_gmf.py
@@ -15,8 +15,6 @@
 
 
 import numpy as np
-import torch
-import torch.nn as nn
 from tqdm.auto import trange
 
 from .pytorch_ncf_base import NCFBase_PyTorch
@@ -136,6 +134,9 @@ class GMF_PyTorch(NCFBase_PyTorch):
         if self.trainable is False:
             return self
 
+        import torch
+        import torch.nn as nn
+
         device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
         self.device = device
         if self.seed is not None:
@@ -144,7 +145,7 @@ class GMF_PyTorch(NCFBase_PyTorch):
             if torch.cuda.is_available():
                 torch.cuda.manual_seed(self.seed)
 
-        from .pytorch_ncf_base import GMF_torch as GMF
+        from .pytorch_models import GMF_torch as GMF
 
         self.model = GMF(
             self.num_users,
@@ -155,12 +156,14 @@ class GMF_PyTorch(NCFBase_PyTorch):
             self.pretrained_GMF,
         ).to(self.device)
 
-        criteria = nn.MSELoss(reduction="sum")
-        optimizer = self.learner(
+        from .pytorch_models import optimizer_dict
+
+        optimizer = optimizer_dict[self.learner](
             self.model.parameters(),
             lr=self.lr,
             weight_decay=self.reg,
         )
+        criteria = nn.MSELoss(reduction="sum")
 
         loop = trange(self.num_epochs, disable=not self.verbose)
         for _ in loop:
@@ -211,6 +214,7 @@ class GMF_PyTorch(NCFBase_PyTorch):
                 raise ScoreException(
                     "Can't make score prediction for (user_id=%d)" % user_idx
                 )
+            import torch
 
             item_ids = torch.from_numpy(np.arange(self.train_set.num_items)).to(
                 self.device
